requirements.txt

flask
flask-sock
ngrok
pyngrok
twilio
python-dotenv
